+++ packages/charts/react-charts/library/src/components/VerticalBarChart/VerticalBarChart.types.ts
@@ -121,7 +121,7 @@ export interface VerticalBarChartProps extends CartesianChartProps {
    * Specifies the mode of the chart.
    * @default 'default'
    */
-  mode?: 'default' | 'plotly';
+  mode?: 'default' | 'plotly' | 'histogram';
 }
 
 /**