+++ packages/charts/react-charts/library/src/components/VerticalStackedBarChart/VerticalStackedBarChart.tsx
@@ -305,7 +305,11 @@ export const VerticalStackedBarChart: React.FunctionComponent<VerticalStackedBar
     _colors = defaultColors;
     _xAxisType = getTypeOfAxis(props.data[0].xAxisPoint, true) as XAxisTypes;
     _lineObject = _getFormattedLineData(props.data);
-    _xAxisInnerPadding = getScalePadding(props.xAxisInnerPadding, props.xAxisPadding, 2 / 3);
+    _xAxisInnerPadding = getScalePadding(
+      props.xAxisInnerPadding,
+      props.xAxisPadding,
+      _xAxisType === XAxisTypes.StringAxis ? 2 / 3 : 1 / 2,
+    );
     _xAxisOuterPadding = getScalePadding(props.xAxisOuterPadding, props.xAxisPadding, 0);
   }
 
@@ -798,9 +802,10 @@ export const VerticalStackedBarChart: React.FunctionComponent<VerticalStackedBar
 
       const xBarScale = d3ScaleLinear()
         .domain(_isRtl ? [xMax, xMin] : [xMin, xMax])
-        .nice()
         .range([_margins.left! + _domainMargin, containerWidth - _margins.right! - _domainMargin]);
-
+      if (!isScalePaddingDefined(props.xAxisInnerPadding, props.xAxisPadding)) {
+        xBarScale.nice();
+      }
       return { xBarScale, yBarScale };
     }
     if (_xAxisType === XAxisTypes.DateAxis) {
@@ -881,16 +886,23 @@ export const VerticalStackedBarChart: React.FunctionComponent<VerticalStackedBar
         let reqWidth = (_xAxisLabels.length + (_xAxisLabels.length - 1) * barGapRate) * barWidth;
         const margin1 = (totalWidth - reqWidth) / 2;
 
-        // Calculate the remaining width after accounting for the space required to render x-axis labels
-        const step = calculateLongestLabelWidth(_xAxisLabels) + 20;
-        reqWidth = (_xAxisLabels.length - _xAxisInnerPadding) * step;
-        const margin2 = (totalWidth - reqWidth) / 2;
+        let margin2 = Number.POSITIVE_INFINITY;
+        if (!props.hideTickOverlap) {
+          // Calculate the remaining width after accounting for the space required to render x-axis labels
+          const step = calculateLongestLabelWidth(_xAxisLabels) + 20;
+          reqWidth = (_xAxisLabels.length - _xAxisInnerPadding) * step;
+          margin2 = (totalWidth - reqWidth) / 2;
+        }
 
         _domainMargin = MIN_DOMAIN_MARGIN + Math.max(0, Math.min(margin1, margin2));
       }
     } else {
       const data = (props.data?.map(point => point.xAxisPoint) as number[] | Date[] | undefined) || [];
-      _barWidth = getBarWidth(props.barWidth, props.maxBarWidth, calculateAppropriateBarWidth(data, totalWidth));
+      _barWidth = getBarWidth(
+        props.barWidth,
+        props.maxBarWidth,
+        calculateAppropriateBarWidth(data, totalWidth, _xAxisInnerPadding),
+      );
       _domainMargin = MIN_DOMAIN_MARGIN + _barWidth / 2;
     }
 
@@ -1197,6 +1209,7 @@ export const VerticalStackedBarChart: React.FunctionComponent<VerticalStackedBar
           xAxisOuterPadding: _xAxisOuterPadding,
         })}
         componentRef={cartesianChartRef}
+        showRoundOffXTickValues={!isScalePaddingDefined(props.xAxisInnerPadding, props.xAxisPadding)}
         /* eslint-disable react/jsx-no-bind */
         children={(props: ChildProps) => {
           return (