+++ packages/react-components/react-menu-grid-preview/library/src/components/MenuGridRowGroupHeader/MenuGridRowGroupHeader.types.ts
@@ -1,5 +1,4 @@
 import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
-import type { MenuGridRowGroupHeaderContextValue } from '../../contexts/menuGridRowGroupHeaderContext';
 
 export type MenuGridRowGroupHeaderSlots = {
   root: Slot<'div'>;
@@ -8,7 +7,3 @@ export type MenuGridRowGroupHeaderSlots = {
 export type MenuGridRowGroupHeaderProps = ComponentProps<MenuGridRowGroupHeaderSlots>;
 
 export type MenuGridRowGroupHeaderState = ComponentState<MenuGridRowGroupHeaderSlots>;
-
-export type MenuGridRowGroupHeaderContextValues = {
-  menuGridRowGroupHeader: MenuGridRowGroupHeaderContextValue;
-};