+++ packages/react-components/react-virtualizer/library/src/utilities/VirtualizerContext/types.ts
@@ -1,6 +1,7 @@
 import * as React from 'react';
 /**
  * {@docCategory Virtualizer}
+ * @deprecated migrated to \@fluentui\-contrib/react\-virtualizer for stable release.
  */
 export type VirtualizerContextProps = {
   contextIndex: number;
@@ -13,5 +14,6 @@ export type VirtualizerContextProps = {
 
 /**
  * Some props are optional on static virtualizer, but required for dynamic.
+ * @deprecated migrated to \@fluentui\-contrib/react\-virtualizer for stable release.
  */
 export type DynamicVirtualizerContextProps = Required<VirtualizerContextProps>;